Splitted classes to folders Beginner and Basic
