+++ src/main/java/Basic/MyArea.java
@@ -1,4 +1,4 @@
-package HomeTasks;
+package Basic;
 
 /*Создать класс MyArea, в нем объявить константу PI = 3.14 и статический метод areaOfCircle,
 который должен принимать радиус и используя PI посчитать площадь круга.*/
