+++ src/main/java/Beginner/ArithmeticAverage.java
@@ -1,4 +1,4 @@
-package HomeTasks;
+package Beginner;
 
 /*Создайте класс ArithmeticAverage.
 Вычислите среднее арифметическое трех целочисленных значений и выведите его на экран.*/
