+++ src/main/java/Beginner/Circle.java
@@ -1,4 +1,4 @@
-package HomeTasks;
+package Beginner;
 
 /*Создайте класс Circle.
 Создайте константу с именем PI (число π «пи»), создайте переменную радиус с именем – r.
