+++ src/main/java/Beginner/Conversion.java
@@ -1,4 +1,4 @@
-package HomeTasks;
+package Beginner;
 
 /*Создайте класс Conversion.
 Напишите программу, которая будет выполнять конвертирование валют. Пользователь вводит:
