+++ src/main/java/Beginner/CustomerArray.java
@@ -1,4 +1,4 @@
-package HomeTasks;
+package Beginner;
 
 import java.util.Arrays;
 import java.util.Scanner;
