+++ src/main/java/Beginner/Fibonacci.java
@@ -1,4 +1,4 @@
-package HomeTasks;
+package Beginner;
 
 /*Напишите метод, возвращающий n-й элемент последовательности Фибоначчи (последовательность Фибоначчи — это последовательность чисел: 0, 1, 1, 2, 3, 5, 8, 13, 21, 34,...
 Следующее число находится путем сложения два числа перед ним.)*/
