+++ src/main/java/Beginner/Interval.java
@@ -1,4 +1,4 @@
-package HomeTasks;
+package Beginner;
 
 /*Создайте класс Interval.
 Напишите программу определения, попадает ли указанное пользователем
