+++ src/main/java/Beginner/OddEven.java
@@ -1,4 +1,4 @@
-package HomeTasks;
+package Beginner;
 
 import java.util.Scanner;
