+++ src/main/java/Beginner/Premium.java
@@ -1,4 +1,4 @@
-package HomeTasks;
+package Beginner;
 
 import java.util.Scanner;
