+++ src/main/java/Beginner/Sort.java
@@ -1,4 +1,4 @@
-package HomeTasks;
+package Beginner;
 
 /*Создайте класс Sort
 В классе создать метод который возвращает массив радомных чисел
