+++ src/main/java/Beginner/Task3.java
@@ -1,4 +1,4 @@
-package HomeTasks;
+package Beginner;
 
 /*Имеется 3 переменные типа int x = 10, y = 12, и z = 3;
         Выполните и рассчитайте результат следующих операций для этих переменных:
