+++ src/main/java/Beginner/VolumeSquare.java
@@ -1,4 +1,4 @@
-package HomeTasks;
+package Beginner;
 
 import java.util.Scanner;
